requirements.txt
python-jose[cryptography] 
passlib[bcrypt]
uvicorn
fastapi
google-generator
bcrypt
python-multipart
